Refactor: check 메서드 TicketCheckServiceImpl 제거
